+++ server/src/main/java/com/krillsson/sysapi/core/domain/history/SystemHistoryEntry.kt
@@ -1,7 +1,5 @@
 package com.krillsson.sysapi.core.domain.history
 
-import com.krillsson.sysapi.core.domain.history.HistoryEntry
-import com.krillsson.sysapi.core.domain.system.SystemLoad
 import java.time.OffsetDateTime
 
-class SystemHistoryEntry(date: OffsetDateTime, value: SystemLoad) : HistoryEntry<SystemLoad>(date, value)
+class SystemHistoryEntry(val date: OffsetDateTime, val value: HistorySystemLoad)