+++ server/src/test/java/com/krillsson/sysapi/core/history/HistoryTest.java
@@ -1,8 +1,10 @@
 package com.krillsson.sysapi.core.history;
 
+import com.krillsson.sysapi.core.domain.history.HistorySystemLoad;
 import com.krillsson.sysapi.util.Clock;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.Mock;
 
 import java.time.OffsetDateTime;
 import java.time.temporal.ChronoUnit;
@@ -13,15 +15,19 @@ import static org.junit.Assert.assertTrue;
 
 public class HistoryTest {
 
-    History<Object> history;
+    HistoryRepository history;
     Clock clock;
+    @Mock
+    HistoryStore store;
+
+    @Mock
+    HistorySystemLoad load;
 
 
     @Before
     public void setUp() throws Exception {
         clock = new Clock();
-        history = new History<>(clock);
-
+        history = new HistoryRepository(clock, store);
     }
 
     @Test
@@ -29,9 +35,9 @@ public class HistoryTest {
 
         OffsetDateTime twoMinutesAgo = OffsetDateTime.now().minusMinutes(2);
         clock.useFixedClockAt(twoMinutesAgo);
-        history.record(new Object());
+        history.record(load);
         clock.useFixedClockAt(twoMinutesAgo.plusMinutes(4));
-        history.record(new Object());
+        history.record(load);
 
         assertThat(history.get().size(), is(2));
         assertTrue(history.get().get(0).getDate().isBefore(history.get().get(1).getDate()));
@@ -46,9 +52,9 @@ public class HistoryTest {
 
         System.out.println(OffsetDateTime.now().plusHours(2).toString());
         clock.useFixedClockAt(twoMinutesAgo);
-        history.record(new Object());
+        history.record(load);
         clock.useFixedClockAt(twoMinutesAgo.plusMinutes(4));
-        history.record(new Object());
+        history.record(load);
 
         assertThat(history.get().size(), is(2));